+++ src/cylinder.c
@@ -6,20 +6,20 @@
 /*   By: tterao <tterao@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/08 14:15:30 by ryhara            #+#    #+#             */
-/*   Updated: 2023/10/11 15:51:16 by tterao           ###   ########.fr       */
+/*   Updated: 2023/10/11 16:13:28 by tterao           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 #include "parse.h"
 
-bool			check_cylinder_cross(float dot, float height, float t);
+bool			check_cylinder_cross(double dot, double height, double t);
 t_cylinder_ray	get_cylinder_ray(t_vector3d start_pos, t_cylinder *cylinder,
 					t_t_data t, t_vector3d camera_ray);
 t_fcolor		get_radiance(t_global_data *data, t_objs *node,
-					t_vector3d ray, const float t);
+					t_vector3d ray, const double t);
 
-static float	_calc_t(t_cylinder *cylinder,
+static double	_calc_t(t_cylinder *cylinder,
 					t_cylinder_ray ray_cylinder, t_t_data t)
 {
 	if (check_cylinder_cross(vector3d_dot(
@@ -31,30 +31,30 @@ static float	_calc_t(t_cylinder *cylinder,
 			cylinder->height, t.t2))
 		t.t = t.t2;
 	else
-		t.t = -1.0f;
+		t.t = -1.0;
 	return (t.t);
 }
 
-float	hit_cylinder(t_vector3d ray, t_vector3d start_pos,
+double	hit_cylinder(t_vector3d ray, t_vector3d start_pos,
 						t_cylinder *cylinder)
 {
 	t_t_data		t;
 	t_cylinder_ray	ray_cylinder;
 
 	t.a = vector3d_mag_sq(vector3d_cross(ray, cylinder->direction));
-	if (t.a == 0.0f)
-		return (-1.0f);
-	t.b = 2.0f * (vector3d_dot(vector3d_cross(ray, cylinder->direction),
+	if (t.a == 0.0)
+		return (-1.0);
+	t.b = 2.0 * (vector3d_dot(vector3d_cross(ray, cylinder->direction),
 				vector3d_cross(vector3d_sub(start_pos, cylinder->coordinate),
 					cylinder->direction)));
 	t.c = vector3d_mag_sq(vector3d_cross(
 				vector3d_sub(start_pos, cylinder->coordinate),
-				cylinder->direction)) - powf(cylinder->radius, 2.0f);
-	t.d = powf(t.b, 2.0f) - (4.0f * t.a * t.c);
+				cylinder->direction)) - powf(cylinder->radius, 2.0);
+	t.d = powf(t.b, 2.0) - (4.0 * t.a * t.c);
 	if (t.d == 0)
-		t.t = (-(t.b) / (2.0f * t.a));
-	t.t1 = (-t.b + sqrtf(powf(t.b, 2.0f) - (4.0f * t.a * t.c))) / (2.0f * t.a);
-	t.t2 = (-t.b - sqrtf(powf(t.b, 2.0f) - (4.0f * t.a * t.c))) / (2.0f * t.a);
+		t.t = (-(t.b) / (2.0 * t.a));
+	t.t1 = (-t.b + sqrtf(powf(t.b, 2.0) - (4.0 * t.a * t.c))) / (2.0 * t.a);
+	t.t2 = (-t.b - sqrtf(powf(t.b, 2.0) - (4.0 * t.a * t.c))) / (2.0 * t.a);
 	ray_cylinder = get_cylinder_ray(start_pos, cylinder, t, ray);
 	return (_calc_t(cylinder, ray_cylinder, t));
 }
@@ -64,7 +64,7 @@ void	render_cylinder_loop(t_global_data *data, t_objs *node)
 	t_vector2d		pos;
 	t_vector3d		camera_ray;
 	t_fcolor		radiance;
-	float			t;
+	double			t;
 
 	pos.y = 0;
 	while (pos.y < WINDOW_HEIGHT)
@@ -74,7 +74,7 @@ void	render_cylinder_loop(t_global_data *data, t_objs *node)
 		{
 			camera_ray = get_camera_ray_dynamic(pos.x, pos.y, data);
 			t = hit_cylinder(camera_ray, data->camera->coordinate, (t_cylinder *)node->obj);
-			if (t < 0.0f)
+			if (t < 0.0)
 				my_mlx_pixel_put(data, pos.x, pos.y, create_rgb(data->background.red, data->background.green, data->background.blue));
 			else
 			{
