+++ src/get_camera_ray.c
@@ -6,7 +6,7 @@
 /*   By: tterao <tterao@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/08 13:59:02 by ryhara            #+#    #+#             */
-/*   Updated: 2023/10/11 15:15:30 by tterao           ###   ########.fr       */
+/*   Updated: 2023/10/11 16:13:28 by tterao           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,13 +31,13 @@ void	screen_data_init(int x, int y, t_global_data *data,
 	screen->screen_unit_x = vector3d_cross(screen->up, data->camera->direction);
 	screen->screen_unit_y = vector3d_cross(data->camera->direction,
 			screen->screen_unit_x);
-	screen->w_s = 2.0f;
+	screen->w_s = 2.0;
 	screen->h_s = WINDOW_HEIGHT * screen->w_s / WINDOW_WIDTH;
 	screen->camera_to_screen = screen->w_s
-		/ (2.0f * (tanf((float)(data->camera->fov * M_PI / 180) / 2.0f)));
-	screen->f_x = (screen->w_s * (float)x) / (WINDOW_WIDTH - 1)
+		/ (2.0 * (tanf((double)(data->camera->fov * M_PI / 180) / 2.0)));
+	screen->f_x = (screen->w_s * (double)x) / (WINDOW_WIDTH - 1)
 		- screen->w_s / 2;
-	screen->f_y = (-screen->h_s * (float)y) / (WINDOW_HEIGHT - 1)
+	screen->f_y = (-screen->h_s * (double)y) / (WINDOW_HEIGHT - 1)
 		+ screen->h_s / 2;
 }
 
@@ -52,8 +52,8 @@ t_vector3d	get_camera_ray(int x, int y, t_global_data *data)
 {
 	t_vector3d	coordinate;
 
-	coordinate.x = ((2.0f * x) / (WINDOW_WIDTH - 1.0f)) - 1.0f;
-	coordinate.y = (-(2.0f * y) / (WINDOW_HEIGHT - 1.0f)) + 1.0f;
+	coordinate.x = ((2.0 * x) / (WINDOW_WIDTH - 1.0)) - 1.0;
+	coordinate.y = (-(2.0 * y) / (WINDOW_HEIGHT - 1.0)) + 1.0;
 	coordinate.z = 0;
 	return (vector3d_sub(coordinate, data->camera->coordinate));
 }
